Split test_state_api into data source client and state manager api tests
