+++ python/ray/tests/BUILD.bazel
@@ -558,6 +558,7 @@ py_test_module_list(
         "test_runtime_env_fork_process.py",
         "test_runtime_env_get_wheel_names.py",
         "test_runtime_env_py_executable.py",
+        "test_state_api_manager.py",
         "test_state_api_summary.py",
         "test_streaming_generator_regression.py",
         "test_system_metrics.py",
@@ -910,6 +911,7 @@ py_test_module_list(
         "test_runtime_env_working_dir_uri.py",
         "test_serialization.py",
         "test_state_api.py",
+        "test_state_api_data_source_client_get.py",
         "test_task_events.py",
         "test_task_events_2.py",
     ],
